src/EMRNext.Core/Services/Clinical/ClinicalWorkflowService.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.Extensions.Logging;
using EMRNext.Core.Models;
using EMRNext.Core.Models.Clinical;

namespace EMRNext.Core.Services.Clinical
{
    public class ClinicalWorkflowService
    {
        // Workflow Status Enumeration
        public enum WorkflowStatus
        {
            Pending,
            InProgress,
            Completed,
            Suspended,
            Cancelled
        }

        // Workflow Priority Levels
        public enum WorkflowPriority
        {
            Low,
            Medium,
            High,
            Urgent
        }

        // Clinical Workflow Model
        public class ClinicalWorkflow
        {
            public Guid Id { get; set; }
            public string PatientId { get; set; }
            public string WorkflowType { get; set; }
            public WorkflowStatus Status { get; set; }
            public WorkflowPriority Priority { get; set; }
            public DateTime CreatedAt { get; set; }
            public DateTime? UpdatedAt { get; set; }
            public string CreatedBy { get; set; }
            public List<WorkflowStep> Steps { get; set; }
            public Dictionary<string, string> Metadata { get; set; }
        }

        // Workflow Step Model
        public class WorkflowStep
        {
            public Guid Id { get; set; }
            public string StepName { get; set; }
            public WorkflowStatus StepStatus { get; set; }
            public DateTime? StartedAt { get; set; }
            public DateTime? CompletedAt { get; set; }
            public string AssignedTo { get; set; }
            public string Description { get; set; }
            public List<WorkflowStepAction> Actions { get; set; }
        }

        // Workflow Step Action Model
        public class WorkflowStepAction
        {
            public Guid Id { get; set; }
            public string ActionName { get; set; }
            public string ActionType { get; set; }
            public DateTime PerformedAt { get; set; }
            public string PerformedBy { get; set; }
            public Dictionary<string, string> ActionDetails { get; set; }
        }

        // Clinical Workflow Management Service
        private readonly ILogger<ClinicalWorkflowService> _logger;
        private readonly IQueryable<Patient> _patientRepository;
        private readonly IQueryable<ClinicalWorkflow> _workflowRepository;

        public ClinicalWorkflowService(
            ILogger<ClinicalWorkflowService> logger,
            IQueryable<Patient> patientRepository,
            IQueryable<ClinicalWorkflow> workflowRepository)
        {
            _logger = logger;
            _patientRepository = patientRepository;
            _workflowRepository = workflowRepository;
        }

        // Create a new clinical workflow
        public async Task<ClinicalWorkflow> CreateWorkflowAsync(
            string patientId, 
            string workflowType, 
            WorkflowPriority priority = WorkflowPriority.Medium)
        {
            // Validate patient exists
            var patient = _patientRepository.FirstOrDefault(p => p.Id == patientId);
            if (patient == null)
            {
                throw new ArgumentException("Patient not found", nameof(patientId));
            }

            var workflow = new ClinicalWorkflow
            {
                Id = Guid.NewGuid(),
                PatientId = patientId,
                WorkflowType = workflowType,
                Status = WorkflowStatus.Pending,
                Priority = priority,
                CreatedAt = DateTime.UtcNow,
                CreatedBy = GetCurrentUserIdentifier(),
                Steps = GenerateDefaultWorkflowSteps(workflowType),
                Metadata = new Dictionary<string, string>()
            };

            // Save workflow
            await SaveWorkflowAsync(workflow);

            _logger.LogInformation(
                $"Created workflow {workflow.Id} for patient {patientId}, type: {workflowType}"
            );

            return workflow;
        }

        // Update workflow status
        public async Task UpdateWorkflowStatusAsync(
            Guid workflowId, 
            WorkflowStatus newStatus)
        {
            var workflow = _workflowRepository
                .FirstOrDefault(w => w.Id == workflowId);

            if (workflow == null)
            {
                throw new ArgumentException("Workflow not found", nameof(workflowId));
            }

            workflow.Status = newStatus;
            workflow.UpdatedAt = DateTime.UtcNow;

            await SaveWorkflowAsync(workflow);

            _logger.LogInformation(
                $"Updated workflow {workflowId} status to {newStatus}"
            );
        }

        // Complete a workflow step
        public async Task CompleteWorkflowStepAsync(
            Guid workflowId, 
            Guid stepId, 
            WorkflowStepAction action)
        {
            var workflow = _workflowRepository
                .FirstOrDefault(w => w.Id == workflowId);

            if (workflow == null)
            {
                throw new ArgumentException("Workflow not found", nameof(workflowId));
            }

            var step = workflow.Steps.FirstOrDefault(s => s.Id == stepId);
            if (step == null)
            {
                throw new ArgumentException("Workflow step not found", nameof(stepId));
            }

            step.StepStatus = WorkflowStatus.Completed;
            step.CompletedAt = DateTime.UtcNow;
            step.Actions ??= new List<WorkflowStepAction>();
            step.Actions.Add(action);

            // Check if all steps are completed
            if (workflow.Steps.All(s => s.StepStatus == WorkflowStatus.Completed))
            {
                workflow.Status = WorkflowStatus.Completed;
            }

            await SaveWorkflowAsync(workflow);

            _logger.LogInformation(
                $"Completed step {stepId} in workflow {workflowId}"
            );
        }

        // Generate default workflow steps based on workflow type
        private List<WorkflowStep> GenerateDefaultWorkflowSteps(string workflowType)
        {
            return workflowType switch
            {
                "MedicalConsultation" => CreateMedicalConsultationSteps(),
                "DiagnosticProcedure" => CreateDiagnosticProcedureSteps(),
                "Prescription" => CreatePrescriptionWorkflowSteps(),
                _ => throw new ArgumentException(
                    $"Unknown workflow type: {workflowType}", 
                    nameof(workflowType))
            };
        }

        // Medical Consultation Workflow Steps
        private List<WorkflowStep> CreateMedicalConsultationSteps()
        {
            return new List<WorkflowStep>
            {
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Patient Intake",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Initial patient registration and vital signs"
                },
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Medical History Review",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Review patient's medical history and current symptoms"
                },
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Physician Consultation",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Physician examination and diagnosis"
                },
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Treatment Plan",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Develop and document treatment plan"
                }
            };
        }

        // Diagnostic Procedure Workflow Steps
        private List<WorkflowStep> CreateDiagnosticProcedureSteps()
        {
            return new List<WorkflowStep>
            {
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Procedure Order",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Physician orders diagnostic procedure"
                },
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Patient Preparation",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Patient preparation and consent"
                },
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Procedure Execution",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Perform diagnostic procedure"
                },
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Result Interpretation",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Analyze and interpret procedure results"
                }
            };
        }

        // Prescription Workflow Steps
        private List<WorkflowStep> CreatePrescriptionWorkflowSteps()
        {
            return new List<WorkflowStep>
            {
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Prescription Order",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Physician creates prescription"
                },
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Pharmacy Verification",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Pharmacy reviews and validates prescription"
                },
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Medication Dispensing",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Prepare and package medication"
                },
                new WorkflowStep
                {
                    Id = Guid.NewGuid(),
                    StepName = "Patient Counseling",
                    StepStatus = WorkflowStatus.Pending,
                    Description = "Provide medication instructions to patient"
                }
            };
        }

        // Save workflow to repository (to be implemented by specific data access layer)
        private async Task SaveWorkflowAsync(ClinicalWorkflow workflow)
        {
            // This would typically involve calling a repository method
            // For now, we'll simulate the save
            _logger.LogInformation($"Saving workflow {workflow.Id}");
            await Task.CompletedTask;
        }

        // Get current user identifier (to be implemented with actual authentication context)
        private string GetCurrentUserIdentifier()
        {
            // In a real implementation, this would come from the authentication context
            return "SYSTEM";
        }
    }
}
